change multiply data type double -> BigDecimal to have really correct results. modify all calls.

+++ src/main/java/Main_chapter_static.java
@@ -1,6 +1,8 @@
 import chapter_static.Calculator;
 import chapter_static.Counters;
 
+import java.math.BigDecimal;
+
 public class Main_chapter_static {
     public static void main(String[] args) {
         Counters myCounters1 = new Counters(0);
@@ -33,7 +35,7 @@ public class Main_chapter_static {
 
         System.out.println("33 + 9 = " + Calculator.add(33, 9));
         System.out.println("75 - 33 = " + Calculator.subtract(75, 33));
-        System.out.println("6 * 7 = " + Calculator.multiply(6, 7));
+        System.out.println("6 * 7 = " + Calculator.multiply(BigDecimal.valueOf(6), BigDecimal.valueOf(7)));
         System.out.println("420 / 10 = " + Calculator.divide(420, 10));
     }
 }
